memsafety.c: fix __INSTR_fsm_list_search
Fix the condition for malloc'ed memory. The problem
was that at this moment, we first remember malloc
using __INSTR_fsm_change_state() and after that we
try to save the size using __INSTR_remember_malloc_size().
__INSTR_remember_malloc_size() calls __INSTR_fsm_list_search(),
that compares id - fsm->id < fsm->size. But the fsm->size == 0
at that moment, so the contidion is false and the automaton
is not found. That means that we do not set the size and
also any other search of this automaton fails.
The proper fix would be to add __INSTR_remember_malloc(),
which we already have in different branch. Before merging
that, we just fix the condition.

Signed-off-by: Marek Chalupa <xchalup4@fi.muni.cz>