Detect also memory that does not make it to the return
If we allocate memory and lost all references to it,
it may not be present in the exit state and we
will not have it in the possiblyLeaked set.
Fix this by storing such allocations into the set
once we detect that the memory may be leaked.
Also, we must move all rules querying the possiblyLeaked set
into a new (later) phase.