analyses: accept a list of lists of analyses
Every list of analyses is parsed and the first one in the list is used,
that is, i.e.:

[["A1", "A2"],
 ["B1", "B2"]]

The analysis A1 is tried and if it fails, A2 is used instead.
Then B1 is tried and if it fails B2 is tried instead.

So the only difference now is that we can say: "run this and that
analysis if the previous one fails".